backend/requirements.txt
pydantic==1.10.12
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
gunicorn==21.2.0
